PatternOS/backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.27.1
python-multipart==0.0.26
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
openai==2.29.0
yfinance==1.2.0
pandas==2.2.1
numpy==1.26.4
pyarrow==16.1.0
ta==0.11.0
mplfinance==0.12.10b0
scipy==1.13.1
apscheduler==3.11.2
python-telegram-bot==21.0
httpx==0.27.0
pytest==8.3.3
